+++ quarkus/addons/messaging/common/src/main/java/org/kie/kogito/addon/quarkus/messaging/endpoint/AbstractQuarkusCloudEventEmitter.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.kie.kogito.addon.quarkus.messaging.common;
+package org.kie.kogito.addon.quarkus.messaging.endpoint;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
