+++ quarkus/addons/messaging/common/src/main/java/org/kie/kogito/addon/quarkus/messaging/rest/QuarkusTopicsInformationResource.java
@@ -16,34 +16,39 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.kie.kogito.addon.messaging.common;
-
-import java.util.List;
+package org.kie.kogito.addon.quarkus.messaging.rest;
 
 import org.kie.kogito.addon.cloudevents.AbstractTopicsInformationResource;
-import org.kie.kogito.event.Topic;
 import org.kie.kogito.event.TopicDiscovery;
 import org.kie.kogito.event.cloudevents.CloudEventMeta;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Component;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
-@RestController
-@RequestMapping("/messaging/topics")
-@Component()
-public class SpringTopicsInformationResource extends AbstractTopicsInformationResource {
+import jakarta.annotation.PostConstruct;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.enterprise.inject.Instance;
+import jakarta.inject.Inject;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
 
-    @Autowired
-    public SpringTopicsInformationResource(TopicDiscovery topicDiscovery, List<CloudEventMeta> cloudEventMetaIterable) {
+@Path("/messaging/topics")
+@ApplicationScoped()
+public class QuarkusTopicsInformationResource extends AbstractTopicsInformationResource {
+
+    @Inject
+    private TopicDiscovery topicDiscovery;
+
+    @Inject
+    private Instance<CloudEventMeta> cloudEventMetaIterable;
+
+    @PostConstruct
+    private void onPostConstruct() {
         setup(topicDiscovery, cloudEventMetaIterable);
     }
 
-    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<Topic>> getTopics() {
-        return ResponseEntity.ok(getTopicList());
+    @GET()
+    @Produces(MediaType.APPLICATION_JSON)
+    public jakarta.ws.rs.core.Response getTopics() {
+        return jakarta.ws.rs.core.Response.ok(getTopicList()).build();
     }
 }
