+++ springboot/addons/messaging/implementation/src/main/java/org/kie/kogito/addon/messaging/config/SpringKafkaConsumerConfig.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.kie.kogito.addon.messaging.common;
+package org.kie.kogito.addon.messaging.config;
 
 import java.util.Map;
